etapa4 finalizada .

+++ routes/articulos.js
@@ -4,7 +4,25 @@ const db = require("../base-orm/sequelize-init");
 const { Op, ValidationError } = require("sequelize");
 
 router.get("/api/articulos", async function (req, res, next) {
-    let items = await db.articulos.findAll({
+  // #swagger.tags = ['Articulos']
+  // #swagger.summary = 'obtiene todos los Articulos'
+  // consulta de articulos con filtros y paginacion
+
+  if (req.query.Pagina) {
+    let where = {};
+    if (req.query.Nombre != undefined && req.query.Nombre !== "") {
+      where.Nombre = {
+        [Op.like]: "%" + req.query.Nombre + "%",
+      };
+    }
+    if (req.query.Activo != undefined && req.query.Activo !== "") {
+      // true o false en el modelo, en base de datos es 1 o 0
+      // convierto el string a booleano
+      where.Activo = (req.query.Activo === 'true'); 
+    }
+    const Pagina = req.query.Pagina ?? 1;
+    const TamañoPagina = 10;
+    const { count, rows } = await db.articulos.findAndCountAll({
       attributes: [
         "IdArticulo",
         "Nombre",
@@ -14,12 +32,36 @@ router.get("/api/articulos", async function (req, res, next) {
         "Activo",
       ],
       order: [["Nombre", "ASC"]],
+      where,
+      offset: (Pagina - 1) * TamañoPagina,
+      limit: TamañoPagina,
+    });
+
+    return res.json({ Items: rows, RegistrosTotal: count });
+    
+  } else {
+    let items = await db.articulos.findAll({
+      attributes: [
+        "IdArticulo",
+        "Nombre",
+        "Precio",
+        "CodigoDeBarra",
+        "IdArticuloFamilia",
+        "Stock",
+        "FechaAlta",
+        "Activo",
+      ],
+      order: [["Nombre", "ASC"]],
     });
     res.json(items);
+  }
 });
 
 router.get("/api/articulos/:id", async function (req, res, next) {
-  let items = await db.articulos.findAll({
+  // #swagger.tags = ['Articulos']
+  // #swagger.summary = 'obtiene un Articulo'
+  // #swagger.parameters['id'] = { description: 'identificador del Articulo...' }
+  let items = await db.articulos.findOne({
     attributes: [
       "IdArticulo",
       "Nombre",
@@ -36,6 +78,13 @@ router.get("/api/articulos/:id", async function (req, res, next) {
 });
 
 router.post("/api/articulos/", async (req, res) => {
+  // #swagger.tags = ['Articulos']
+  // #swagger.summary = 'agrega un Articulos'
+  /*    #swagger.parameters['item'] = {
+                in: 'body',
+                description: 'nuevo Articulo',
+                schema: { $ref: '#/definitions/Articulos' }
+    } */
   try {
     let data = await db.articulos.create({
       Nombre: req.body.Nombre,
@@ -48,30 +97,30 @@ router.post("/api/articulos/", async (req, res) => {
     });
     res.status(200).json(data.dataValues); // devolvemos el registro agregado!
   } catch (err) {
-    const messages = {};
     if (err instanceof ValidationError) {
-      err.errors.forEach((error) => {
-        let message;
-        switch (error.validatorKey) {
-          case "not_unique":
-            message = (error.value ?? 'el valor de este campo') + " ya existe en la tabla";
-            error.path = error.path.replace("_UNIQUE", "");
-            break;
-          default:
-            message = error.message;  // msg del modelo o por defecto
-        }
-        messages[error.path] = message;
-      });
-      res.status(400).json(messages);
-    } else throw err; // desencadeno el mismo error inicial (error desconocido)
+      // si son errores de validacion, los devolvemos
+      let messages = '';
+      err.errors.forEach((x) => messages += (x.path ?? 'campo') + ": " + x.message + '\n');
+      res.status(400).json({message : messages});
+    } else {
+      // si son errores desconocidos, los dejamos que los controle el middleware de errores
+      throw err;
+    }
   }
 });
 
 router.put("/api/articulos/:id", async (req, res) => {
+  // #swagger.tags = ['Articulos']
+  // #swagger.summary = 'actualiza un Articulo'
+  // #swagger.parameters['id'] = { description: 'identificador del Articulo...' }
+  /*    #swagger.parameters['Articulo'] = {
+                in: 'body',
+                description: 'Articulo a actualizar',
+                schema: { $ref: '#/definitions/Articulos' }
+    } */
 
-    try {
-      let data = await db.articulos
-    .update(
+  try {
+    let data = await db.articulos.update(
       {
         Nombre: req.body.Nombre,
         Precio: req.body.Precio,
@@ -83,33 +132,54 @@ router.put("/api/articulos/:id", async (req, res) => {
       },
       { where: { IdArticulo: req.params.id } }
     );
-    res.status(200).json(data.dataValues); // devolvemos el registro modificado!
+    res.sendStatus(200);
   } catch (err) {
-    const messages = {};
     if (err instanceof ValidationError) {
-      err.errors.forEach((error) => {
-        let message;
-        switch (error.validatorKey) {
-          case "not_unique":
-            message = (error.value ?? 'el valor de este campo') + " ya existe en la tabla";
-            error.path = error.path.replace("_UNIQUE", "");
-            break;
-          default:
-            message = error.message;  // msg del modelo o por defecto
-        }
-        messages[error.path] = message;
-      });
-      res.status(400).json(messages);
-    } else throw err; // desencadeno el mismo error inicial (error desconocido)
+      // si son errores de validacion, los devolvemos
+      let messages = '';
+      err.errors.forEach((x) => messages += x.path + ": " + x.message + '\n');
+      res.status(400).json({message : messages});
+    } else {
+      // si son errores desconocidos, los dejamos que los controle el middleware de errores
+      throw err;
+    }
   }
 });
 
 router.delete("/api/articulos/:id", async (req, res) => {
-  let data = await db.articulos.destroy({
-    where: { IdArticulo: req.params.id },
-  });
-  if (data==1) res.sendStatus(200);
-  else res.sendStatus(404);
-});
+  // #swagger.tags = ['Articulos']
+  // #swagger.summary = 'elimina un Articulo'
+  // #swagger.parameters['id'] = { description: 'identificador del Articulo..' }
 
+  let bajaFisica = false;
+
+  if (bajaFisica) {
+    // baja fisica
+    let filasBorradas = await db.articulos.destroy({
+      where: { IdArticulo: req.params.id },
+    });
+    if (filasBorradas == 1) res.sendStatus(200);
+    else res.sendStatus(404);
+  } else {
+    // baja logica
+    try {
+      let data = await db.sequelize.query(
+        "UPDATE articulos SET Activo = case when Activo = 1 then 0 else 1 end WHERE IdArticulo = :IdArticulo",
+        {
+          replacements: { IdArticulo: +req.params.id },
+        }
+      );
+      res.sendStatus(200);
+    } catch (err) {
+      if (err instanceof ValidationError) {
+        // si son errores de validacion, los devolvemos
+        const messages = err.errors.map((x) => x.message);
+        res.status(400).json(messages);
+      } else {
+        // si son errores desconocidos, los dejamos que los controle el middleware de errores
+        throw err;
+      }
+    }
+  }
+});
 module.exports = router;
